detectado bug na hora de incrementar o movimento horizontal/vertical, falta arrumar ><

+++ Event.h
@@ -17,7 +17,7 @@
 #include <ctime>        //std::time_t
 #include <time.h>       //time_t, time, ctime
 #include <string>       //std::string
-
+#include "MessageDealer.h"
 class Event {
 public:
     Event(std::string id, std::string video_path, std::string video_name, std::string video_extention, std::time_t start_time);
@@ -29,18 +29,18 @@ public:
     std::string getVideoName();
     std::string getVideoFullName();
     std::time_t getStartTime();
-    long getDuration();
+    double getDuration();
     long getFramesQuantity();
     long getHorizontalDirection();
     long getVerticalDirection();
     double getMotionQuantity();
     
-    void setDuration(long duration); // define a duração em segundos do evento
+    void setDuration(double duration); // define a duração em segundos do evento
     void setFramesQuantity(long quantity); // define a quantidade acumulada de frames (quadros de vídeo) do evento
     void setMotionQuantity(double quantity); // define a quantidade de movimentação acumulada do evento
     void incrementFramesQuantity(); // incrementa de um a quantidade acumulada de frames
     void incrementMotionQuantity(double quantity); // incrementa a quantidade acumulada de movimentação do evento com a quantidade passada
-    void incrementHorizontalDirection(short int quantity); // incrementar de 1 significa incremento para a direita, e -1, para a esquerda;
+    void incrementHorizontalDirection(int quantity); // incrementar de 1 significa incremento para a direita, e -1, para a esquerda;
     void incrementVerticalDirection(short int quantity); // incrementar de 1 significa incremento para cima, e -1, para baixo;
 private:
     std::string _id; // identificação do evento
@@ -48,9 +48,9 @@ private:
     std::string _videoName;
     std::string _videoExtention; // extensão usada para salvar o vídeo do evento (Ex: .mp4, .avi, etc...)
     std::time_t _startTime;
-    long _duration; // duração do evento em segundos;
+    double _duration; // duração do evento em segundos;
     long _framesQuantity; // quantidade acumulada de frames (quadros de vídeo) do evento
-    long _horizontalDirection; // valores negativos representam direção média de movimentos para a esquerda, e positivos para a direita;
+    int _horizontalDirection; // valores negativos representam direção média de movimentos para a esquerda, e positivos para a direita;
     long _verticalDirection; // valores negativos representam direção média de movimentos para baixo, e positivos para cima;
     double _motionQuantity; // quantia acumulada de movimentação detectada do evento
